+++ src/traces/sankey/constants.js
@@ -1,0 +1,20 @@
+/**
+* Copyright 2012-2017, Plotly, Inc.
+* All rights reserved.
+*
+* This source code is licensed under the MIT license found in the
+* LICENSE file in the root directory of this source tree.
+*/
+
+'use strict';
+
+module.exports = {
+    nodeTextOffsetHorizontal: 4,
+    nodeTextOffsetVertical: 3,
+    nodePadAcross: 10,
+    sankeyIterations: 50,
+    forceIterations: 5,
+    forceTicksPerFrame: 10,
+    duration: 500,
+    ease: 'linear'
+};