+++ test/jasmine/tests/mapbox_test.js
@@ -709,6 +709,22 @@ describe('@noCI, mapbox plots', function() {
         assertMouseMove(blankPos, 0).then(function() {
             return assertMouseMove(pointPos, 1);
         })
+        .then(function() {
+            return Plotly.restyle(gd, {
+                'hoverlabel.bgcolor': 'yellow',
+                'hoverlabel.font.size': [[20, 10, 30]]
+            });
+        })
+        .then(function() {
+            return assertMouseMove(pointPos, 1);
+        })
+        .then(function() {
+            var path = d3.select('g.hovertext').select('path');
+            var text = d3.select('g.hovertext').select('text.nums');
+
+            expect(path.style('fill')).toEqual('rgb(255, 255, 0)', 'bgcolor');
+            expect(text.style('font-size')).toEqual('20px', 'font.size[0]');
+        })
         .catch(failTest)
         .then(done);
     }, LONG_TIMEOUT_INTERVAL);